Add wiki: EverybodyWiki
"EverybodyWiki is an inclusionist online encyclopedia where anyone can write about anything!"

The straw that broke the camel's back and made me finally start this project:
https://en.everybodywiki.com/List_of_most_starred_GitHub_repositories

I'd never heard of this wiki before, and it just reminded me for the last time,
how awesome it would be to just have a place where all the wikis are collected,
for easier and sooner discovery and maybe also encouraging the creation of
more of them!

Most people (like myself a while ago) don't know that Wikipedia itself is a wiki,
and that wiki software is actually a thing in and of itself. It's just a magical thing
when you realize that there is a whole world of wikis out there that you may not
have even been aware of before!! And that, ladies and gentlemen, is something
worth sharing.